cambios finales

+++ backend/api-Principal/src/main/java/com/example/ecommerce/Dto/ProductoDto.java
@@ -1,10 +1,10 @@
 package com.example.ecommerce.Dto;
 
-import com.example.ecommerce.Dto.Request.ProductoRequestDto;
+
 import com.example.ecommerce.Model.Producto;
 
 import lombok.Getter;
-import lombok.NoArgsConstructor;
+
 import lombok.Setter;
 
 @Getter
@@ -14,7 +14,7 @@ public class ProductoDto {
     private Long id;
     private String nombre;
     private String modelo;
-    private Double precio;
+    private int precio;
     private String descripcion;
     private Long categoria_id;
     private Integer cantidad;
